requirements.txt
fastapi
uvicorn
playwright
firebase-admin
python-dotenv
beautifulsoup4
newspaper3k